Add files via upload
This file examines the data of Titanic passengers in order to comprehend the survivors of the catastrophe. It initializes the data by filling in the data that is missing ages among other data. The review illustrates that in general, 38 per cent of the passengers survived. The survival rates of women were substantially higher in comparison to men, the first-class passengers faced better survival opportunities than third-class, and the place of boarding the ship impacted on the survival rates. The file also represents these patterns in charts, ensuring that the women and children first rule was not imaginary and that wealth and social status contributed immensely to the survivor of the tragedy.